+++ srv/service.ts
@@ -113,6 +113,7 @@ export class AdminService extends cds.ApplicationService {
         this.on('syncMatchResults', this.adminHandler.syncMatchResults.bind(this.adminHandler));
         this.on('lockMatchBetting', this.adminHandler.lockMatchBetting.bind(this.adminHandler));
         this.on('lockTournamentBetting', this.adminHandler.lockTournamentBetting.bind(this.adminHandler));
+        this.on('setPlayerScoreBetsProcessed', this.adminHandler.setPlayerScoreBetsProcessed.bind(this.adminHandler));
         this.on('getAvailableCompetitions', this.adminHandler.getAvailableCompetitions.bind(this.adminHandler));
         this.on('importTournament', this.adminHandler.importTournament.bind(this.adminHandler));
         this.on('provisionSandboxUsers', this.adminHandler.provisionSandboxUsers.bind(this.adminHandler));